+++ src/classes/api/index.ts
@@ -16,14 +16,14 @@ import {
     PredefinedCalendars,
     PresetTimeOfDay,
     TimeKeeperStatus,
-    YearNamingRules
+    YearNamingRules, NoteRepeat
 } from "../../constants";
 import PF2E from "../systems/pf2e";
 import {AdvanceTimeToPreset, DateToTimestamp, FormatDateTime, TimestampToDate} from "../utilities/date-time";
 import DateSelectorManager from "../date-selector/date-selector-manager"
 import PredefinedCalendar from "../configuration/predefined-calendar";
 import Renderer from "../renderer";
-import {MainApplication, CalManager, SC} from "../index";
+import {MainApplication, CalManager, SC, NManager} from "../index";
 import {canUser} from "../utilities/permissions";
 
 /**
@@ -63,6 +63,10 @@ export {YearNamingRules};
  * The sun positions
  */
 export {PresetTimeOfDay};
+/**
+ * The options for note repeating
+ */
+export{NoteRepeat}
 
 /**
  * This function is used to activate event listeners for calendars displayed with the [sc-full-calendar](#sc-full-calendar) Handlebar helper.
@@ -79,6 +83,43 @@ export {PresetTimeOfDay};
  */
 export function activateFullCalendarListeners(calendarId: string, onMonthChange: Function | null = null, onDayClick: Function | null = null): void{
     Renderer.CalendarFull.ActivateListeners(calendarId, onMonthChange, onDayClick);
+}
+
+/**
+ * This function adds a new note to the calendar
+ *
+ * @param title The title of the new note
+ * @param content The contents of the new note
+ * @param starDate The date and time the note starts
+ * @param endDate The date and time the note ends (can be the same as the start date)
+ * @param allDay If the note lasts all day or if it has a specific time duration. Whether to ignore the time portion of the start and end dates.
+ * @param repeats If the note repeats and how often it does
+ * @param categories A list of note categories to assign to this note
+ * @param calendarId Optional parameter to specify the ID of the calendar to add the note too. If not provided the current active calendar will be used.
+ *
+ * @returns The newly created JournalEntry that contains the note data, or null if there was an error encountered.
+ *
+ * @example
+ * ```javascript
+ * const newJournal = await SimpleCalendar.api.addNote('Christmas Day','Presents!', {year: 2022, month: 11, day: 24, hour: 0, minute: 0, seconds: 0}, {year: 2022, month: 11, day: 24, hour: 0, minute: 0, seconds: 0}, true, SimpleCalendar.api.NoteRepeat.Yearly, ['Holiday']);
+ * // Will create a new note on Christmas day of 2022 that lasts all day and repeats yearly.
+ * ```
+ */
+export async function addNote(title: string, content: string, starDate: SimpleCalendar.DateTime, endDate: SimpleCalendar.DateTime, allDay: boolean, repeats: NoteRepeat, categories: string[], calendarId: string = 'active'): Promise<StoredDocument<JournalEntry> | null>{
+    const activeCalendar = calendarId === 'active'? CalManager.getActiveCalendar() : CalManager.getCalendar(calendarId);
+    if(activeCalendar){
+        return await NManager.createNote( title, content, {
+                calendarId: activeCalendar.id,
+                startDate: starDate,
+                endDate: endDate,
+                allDay: allDay,
+                order: 0,
+                categories: categories,
+                repeats: repeats,
+                remindUsers: []
+            }, activeCalendar, false );
+    }
+    return null;
 }
 
 /**
@@ -1086,6 +1127,44 @@ export function getLeapYearConfiguration(calendarId: string = 'active'): SimpleC
         Logger.error(`SimpleCalendar.api.getLeapYearConfiguration - Unable to find a calendar with the passed in ID of "${calendarId}"`);
     }
     return null;
+}
+
+/**
+ * Gets all notes that the current user is able to see for the specified calendar.
+ *
+ * @param calendarId Optional parameter to specify the ID of the calendar to get the notes from. If not provided the current active calendar will be used.
+ *
+ * @returns A list of Journal entries that contain the note data.
+ *
+ * @example
+ * ```javascript
+ * // Returns an array of [JournalEntry](https://foundryvtt.com/api/JournalEntry.html) objects
+ * SimpleCalendar.api.getNotes();
+ *```
+ */
+export function getNotes(calendarId: string = 'active'): (StoredDocument<JournalEntry> | undefined)[]{
+    calendarId = calendarId === 'active'? CalManager.getActiveCalendar().id : calendarId;
+    return NManager.getNotes(calendarId).map(n => (<Game>game).journal?.get(n.entryId));
+}
+
+/**
+ * Gets all notes that the current user is able to see for the specified date from the specified calendar.
+ * @param year The year of the date to get the notes for.
+ * @param month The index of the month to get the notes for.
+ * @param day The index of the day to get the notes for.
+ * @param calendarId Optional parameter to specify the ID of the calendar to get the notes from. If not provided the current active calendar will be used.
+ *
+ * @returns A list of Journal entries that contain the note data.
+ *
+ * @example
+ * ```javascript
+ * // Returns an array of [JournalEntry](https://foundryvtt.com/api/JournalEntry.html) objects
+ * SimpleCalendar.api.getNotesForDay(2022, 11, 24);
+ * ```
+ */
+export function getNotesForDay(year: number, month: number, day: number, calendarId: string = 'active'): (StoredDocument<JournalEntry> | undefined)[]{
+    calendarId = calendarId === 'active'? CalManager.getActiveCalendar().id : calendarId;
+    return NManager.getNotesForDay(calendarId, year, month, day).map(n => (<Game>game).journal?.get(n.entryId));
 }
 
 /**